requirements_gpu.txt
tensorflow-gpu==1.15.4
scikit-learn
scipy
numpy
pandas
tqdm
psutil